requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
fastapi==0.108.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-fragments==1.2.1
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytailwindcss==0.2.0
pytest==7.4.4
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.32.0.post1
tinydb==4.8.0
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
requests==2.31.0
celery==5.3.6
flower==2.0.1
redis==5.0.1
aioredis==2.0.1